rename RemoveAllAssociatedProductCategoriesCommand/Handler
